requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm==0.9.16
pillow>=10.0.0